+++ src/components/NewFeatures/NewFeatures.js
@@ -19,6 +19,24 @@ const NewFeatures = () => {
         </div>
       </HeaderSection>
       <section className="dp-container">
+        <div className="dp-block-wlp dp-box-shadow m-t-36">
+          <div style={{ marginLeft: '20px' }}>
+            <h3>Compra de planes</h3>
+            <Link to={'/plan-selection?promo-code=ALLPLANS'}>Ir a Comprar Plan</Link>
+          </div>
+        </div>
+        <div className="dp-block-wlp dp-box-shadow m-t-36 m-b-36">
+          <div style={{ marginLeft: '20px' }}>
+            <h3>Plan Calculator</h3>
+            <Link
+              to={
+                '/plan-selection/standard/subscribers?selected-plan=18&advanced-pay=year&promo-code=ALLPLANS'
+              }
+            >
+              Ir a calculadora de planes
+            </Link>
+          </div>
+        </div>
         <div className="dp-block-wlp dp-box-shadow">
           <div style={{ marginLeft: '20px' }}>
             <h3>Maestro de Suscriptores</h3>
@@ -39,28 +57,8 @@ const NewFeatures = () => {
         </div>
         <div className="dp-block-wlp dp-box-shadow m-t-36">
           <div style={{ marginLeft: '20px' }}>
-            <h3>Compra de planes</h3>
-            <Link to={'/plan-selection?advanced-pay=year&promo-code=ALLPLANS&planId=18'}>
-              Ir a Comprar Plan
-            </Link>
-          </div>
-        </div>
-        <div className="dp-block-wlp dp-box-shadow m-t-36">
-          <div style={{ marginLeft: '20px' }}>
             <h3>Formulario de Agencias</h3>
             <Link to={'/email-marketing-agencies'}>Ir a Formulario de Agencias</Link>
-          </div>
-        </div>
-        <div className="dp-block-wlp dp-box-shadow m-t-36 m-b-36">
-          <div style={{ marginLeft: '20px' }}>
-            <h3>Plan Calculator</h3>
-            <Link
-              to={
-                '/plan-selection/standard-subscribers?selected-plan=18&advanced-pay=year&promo-code=ALLPLANS'
-              }
-            >
-              Ir a calculadora de planes
-            </Link>
           </div>
         </div>
         <div className="dp-block-wlp dp-box-shadow m-t-36">
